+++ Example/Tests/GiniVisionDocumentValidatorTests.swift
@@ -1,20 +1,17 @@
 //
-//  GINIVisionDocumentTests.swift
+//  GiniVisionDocumentValidatorTests.swift
 //  GiniVision_Tests
 //
-//  Created by Enrique del Pozo Gómez on 8/30/17.
-//  Copyright © 2017 Gini GmbH. All rights reserved.
+//  Created by Enrique del Pozo Gómez on 4/13/18.
+//  Copyright © 2018 Gini GmbH. All rights reserved.
 //
 
 import XCTest
 @testable import GiniVision
 
-class GINIVisionDocumentTests: XCTestCase {
+final class GiniVisionDocumentValidatorTests: XCTestCase {
     
-    let giniConfiguration: GiniConfiguration = GiniConfiguration()
-    var filePickerManager: DocumentPickerCoordinator {
-        return DocumentPickerCoordinator()
-    }
+    let giniConfiguration = GiniConfiguration()
     
     func testExcedeedMaxFileSize() {
         let higherThan10MBData = generateFakeData(megaBytes: 12)
@@ -24,8 +21,8 @@ class GINIVisionDocumentTests: XCTestCase {
         XCTAssertThrowsError(try GiniVisionDocumentValidator.validate(pdfDocument,
                                                                       withConfig: giniConfiguration),
                              "Files with a size lower than 10MB should be valid") { error in
-            XCTAssert(error as? DocumentValidationError == DocumentValidationError.exceededMaxFileSize,
-                      "should indicate that max file size has been exceeded")
+                                XCTAssert(error as? DocumentValidationError == DocumentValidationError.exceededMaxFileSize,
+                                          "should indicate that max file size has been exceeded")
         }
     }
     
@@ -33,12 +30,12 @@ class GINIVisionDocumentTests: XCTestCase {
         let lowerThanOrEqualTo10MBData = generateFakeData(megaBytes: 10)
         
         let pdfDocument = GiniPDFDocument(data: lowerThanOrEqualTo10MBData)
-
+        
         XCTAssertThrowsError(try GiniVisionDocumentValidator.validate(pdfDocument,
                                                                       withConfig: giniConfiguration),
                              "Files with a size greater than 10MB should not be valid") { error in
-            XCTAssert(error as? DocumentValidationError != DocumentValidationError.exceededMaxFileSize,
-                      "should indicate that max file size has been exceeded")
+                                XCTAssert(error as? DocumentValidationError != DocumentValidationError.exceededMaxFileSize,
+                                          "should indicate that max file size has been exceeded")
         }
     }
     
@@ -55,4 +52,5 @@ class GINIVisionDocumentTests: XCTestCase {
         let length = lengthInMB * 1000000
         return Data(count: length)
     }
+    
 }
